+++ simulation.cc
@@ -2,13 +2,13 @@
 
 #include <glm/gtx/compatibility.hpp>
 
-static void interpolate_entities(const entity &s1, const entity &s2, double t, entity *out)
+static void interpolate_entities(const entity &s1, const entity &s2, float t, entity *out)
 {
-	out->position = glm::lerp(s1.position, s2.position, (float)t);
-	out->angle =    glm::lerp(s1.angle,    s2.angle,    (float)t);
+	out->position = glm::lerp(s1.position, s2.position, t);
+	out->angle =    glm::lerp(s1.angle,    s2.angle,    t);
 }
 
-void interpolate_states(const state &s1, const state &s2, double t, state *out)
+void interpolate_states(const state &s1, const state &s2, float t, state *out)
 {
 	/* not accounting for removed entities */
 
@@ -20,9 +20,9 @@ void interpolate_states(const state &s1, const state &s2, double t, state *out)
 		interpolate_entities(s1.entities[i], s2.entities[i], t, &out->entities[i]);
 }
 
-void state::update(double t, double dt)
+void state::update(float t, float dt)
 {
 	for (size_t i = 0; i < entities.size(); ++i)
-		entities[i].angle += 5.f * (float)dt;
+		entities[i].angle += 5.f * dt;
 }
